Add pulseScript file
This file is the adapter between user expression and the envelope math function.
